+++ Front-end/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "GerenciadorUniversitario",
+  "name": "Front-end",
   "version": "0.1.0",
   "private": true,
   "devDependencies": {
